Pin a case to compare
Now we can fix a case in the table comparator using the pin icon that appears with the case id. This button will create a strip chart with the case similarity with all the casebase and update the table. Additionally, the button is hidden if the itemId does not exist