Heatmap
The heatmap class has been translated to Typescript. It extends the PlotObserver class so the controller has been also modified.
TODO: The new class must be documented.

HACK: Plotly type file misses the orientation for the ColorBar so this line must be included in the ColorBar definition:

LINE 1537:
orientation: "v" | "h";