First steps on taxonomy visualization
We start taxonomy visualization using vis.js and a taxonomy defined for car manufacturer. Car dataset now includes a taxonomy, that is visualized on the bottom of the page when car dataset is loaded. Two taxonomy instances are selected as an example (selected from the controller)
Casebases can use a taxonomy so casebaseDAO has been extended for supporting this feature.
We added a new data type for taxonomies and some common operations and a visual component for visualizing the taxonomy trees.