Select cases based on taxonomy concepts
We can click on a concept on the general taxonomy view and a case that contains this label is selected on the case comparator. Additionally, the heatmap is transformed into a stripe chart for the selected case.
The taxonomy viewer and the controller have been modified to implement this feature. Additionally, casebaseDAO now caches the cases indexed by taxonomy labels and it implements a new method to randomly select a case based on a taxonomy label.