Enhancing heatmap visualization using hierarchical clustering
We changed the way that the items are arranged in the heatmap: we use a hierarchical clustering algorithm to sort the items and visualize them. This way, we can see relationships among sets of cases.
To do that, we add a new dependency and a "util" script for sorting using the HC algorithm, and we modified the heatmap to use is and enhance the visualization